Extract tests for isIntlDateTimeFormat to own file
